+++ Modules/Functions.py
@@ -84,22 +84,3 @@ def About_Window ():
     L5.place (x = 20, y = 320)
 
     About.mainloop ()
-
-def Editor ():
-
-    messagebox.showwarning ("Editor warning", "Warning! Editor is not ready, right now only the ability to write something is available without the ability to save to a file!")
-
-    Editor_Window = tk.Tk ()
-    Editor_Window.configure (bg = 'white')
-    Editor_Window.title ("Factorial Editor")
-    Editor_Window.geometry ("1000x650")
-
-    Text_Box = tk.Text (Editor_Window, width = 100, font = ("Montserrat Bold", 12))
-    Text_Box.pack ()
-
-    Save_button = ttk.Button (Editor_Window, text = "Save")
-    Save_button.config (state = "disabled")
-
-    Save_button.place (x = 850, y = 550, width = 130, height = 75)
-
-    Editor_Window.mainloop ()
